Probably fixed rotaion bug

+++ src/main/java/zombienado_v1/server/serverUnits/ServerUnit.java
@@ -7,7 +7,7 @@ public interface ServerUnit {
 
     int getX();
     int getY();
-    double getRotation();
+    float getRotation();
     int getID();
 
 }
